refactor(web): split WeaponSection into folder with sub-components
